Moving test to logical place
