+++ src/cert/ext.rs
@@ -1015,3 +1015,33 @@ pub mod oid {
     pub const PE_SUBJECT_INFO_ACCESS: Oid<&[u8]>
         = Oid(&[43, 6, 1, 5, 5, 7, 1, 11]);
 }
+
+
+//------------ Tests ---------------------------------------------------------
+
+// is pub so that we can use a parsed test IdCert for now for testing
+#[cfg(test)]
+pub mod tests {
+
+    use super::*;
+    use ber::encode::Values;
+
+    #[test]
+    fn should_encode_basic_ca() {
+        let ba = BasicCa::new(true, true);
+        let mut v = Vec::new();
+        ba.encode().write_encoded(Mode::Der, &mut v).unwrap();
+
+        // 48 15            Sequence with length 15
+        //  6 3 85 29 19       OID 2.5.29.19 basicConstraints
+        //  1 1 255              Boolean true
+        //  4 5                OctetString of length 5
+        //     48 3               Sequence with length 3
+        //        1 1 255           Boolean true
+
+        assert_eq!(
+            vec![48, 15, 6, 3, 85, 29, 19, 1, 1, 255, 4, 5, 48, 3, 1, 1, 255 ],
+            v
+        );
+    }
+}
